+++ src/Presenter/DeliveryTimePresenter.php
@@ -61,7 +61,7 @@ class DeliveryTimePresenter
         );
 
         return $this->context->smarty->fetch(
-            $this->module->getLocalPath().'/views/templates/hook/front/carrier-delivery-time.tpl'
+             'module:dpdbaltics/views/templates/hook/front/carrier-delivery-time.tpl'
         );
     }
 }