+++ CHANGELOG.md
@@ -173,5 +173,5 @@
 ## [3.2.21]
 - Fixed label printing error and 0 goods price displaying
 - Improved price calculation
-- Carriers error handling improvements
 - Fixed issues with carrier shown as "Free" when it is not needed to display
+- Carriers error handling improvements